Delay the time between: 1)the instant we detect the current index changed in a tree view and 2) the instant we signal to the outside the fact that the current index changed.
The reason for this is that otherwise we might signal a current index
change while the models didn't stabilize yet. So delaying this signal we
know it'll be applied when it's "quiet".

We previously had a crash on context remove. The ActionListModel was
changing in between beginRemoveRows() and endRemoveRows() causing
corruption in the new model. This commit directly address this.

Thanks to Alexis and Olivier for investigating this one!

CCMAIL: menard@kde.org
CCMAIL: ogoffart@kde.org