Don't display "out of focus" items.
Currently the SelectionProxyModel has to accept the rows of the parents
of currently selected items in the side bar (mostly for technical
reasons). So modify this proxy model to set the size hint of such items
to the null QSize, now in such cases, the latest model in the proxy
chain just remove those lines in order to have a clean segmented list
displayed to the user.