Label projects and categories "parent: child".
Add another proxy which changes the display string of the groups in the
action list editor (use the opportunity to rename the other proxies to a
"Group" naming). This new proxy change the display role to something
like "parent: son: grandson: ...". Handier to read for the user in long
lists.