Build models with the extra Zanshin metadata
Modify the expand* implementations of StandardModelBuilderBehavior to
also include the ItemType, Uid and ParentUid roles to the created
QStandardItems in order to simulate the behavior of Zanshin's
TodoMetadataModel from our tests.

An extra tag is added to declare todos which are referenced by other
todos within the same collection (ReferencedTag), in their metadata
they'll also get the ProjectTogo ItemType.

This part of the standard builder behavior is optional and can be turned
off (to test TodoMetadataModel itself).

Signed-off-by: Kevin Ottens <ervin@kde.org>