Refactor to avoid ETM::modelIndexesForItem
In order to make our model properly testable, we should avoid using
Akonadi::EntityTreeModel::modelIndexesForItem as this one expect an
Akonadi::Session to be available which we want to avoid for our tests.

So make TodoCategoriesModel (resp. TodoTreeModel) use methods in
CategoryManager (resp. TodoHelpers) which do the mapping properly for
us (or when applicable avoid the mapping altogether).