Add a Zanshin plugin for Kontact
Added a ReadOnlyPart and a KontactInterface::Plugin to integrate
Zanshin directly in the Kontact window. Also tune the ActionListEditor
so that it doesn't set a shortcut for the "New Action" action when the
"ZANSHIN_KONTACT_PLUGIN" environment variable is set. This variable is
set only from the plugin before the part (and then MainComponent) is
created.