Make the popup resize logic of the combo box reusable
In showPopup() we readjust the popup size and position to make it easier
to read and avoid having it bound to the combo box width. That logic
has been factored out into a method taking a geometry in, and the ideal
geometry out.

Obviously now showPopup() uses this new method.