Get a larger completer view
Reuse finalizePopupGeometry() from the combo box in order to determine
the proper geometry for completion. Obviously makes
ActionListCompleterView tied to ActionListComboBox, but they're meant to
be always used together anyway.

The geometry is enforced from within resizeEvent() which is the last
step before the completion popup is actually shown, so we use a nasty
static flag to avoid infinite recursion.