Fix crash when a Project with no marker disappears
If the project selected in the SideBar is the same than the top project
in the ActionListEditor, the selection in SideBar must change because
the selected item is no longer present in the list.

To Change the selection we react on the remove row in
SelectionProxyModel and if the m_selectedRows is empty because the
project has disappeared from the SideBar, in this case we force the
selection of parent.