TodoMetadataModel now inherits from KIdentityProxyModel
KIdentityProxyModel is a much better base than QSortFilterProxyModel,
as TodoMetadataModel just adds a couple of roles and makes no
structural modification whatsoever. It's also simpler (don't need to
maintain a very complex cache), and as a matter of fact changing this
base class fixes the crash we have when we drag and drop projects
between collections (as it goes through "item moved" signals).