Workaround Qt 4.8 behavior change of QAbstractProxyModel
OK, that one was tough! Turns out that in Qt 4.8 the behavior of
QAbstractProxyModel changed because it now provides a reimplementation
of hasChildren() not based on rowCount() & columnCount(), but by
forwarding the request to the source model (which in my opinion doesn't
really make sense...).

Because of that, all the proxies or views on top of subclasses of
TodoProxyModelBase were getting totally confused because of
contradictions between hasChildren() and rowCount() results.

That explains the random crashes, or empty views experienced by users
running with Qt 4.8 beta.

Pfew! I'm glad I finally noticed those reports were mentionning Qt 4.8.

BUG: 283066
BUG: 283067