Really repair the move dialog when we move inbox items to collections
The check in moveTodoToProject (first implementation) was too
restrictive, and didn't assume the parent id could be empty (which is
the case for moving something from inbox to a collection). Beside, it
was clearly duplicating code from moveTodoToProject (second
implementation), so remove the check altogether, it'll be verified later
on anyway.

CCMAIL: mbensi@ipsquad.net