Don't forget to reparent promoted todos from the inbox
When a todo get promoted into a project, it is just a dataChanged, and
its parent id didn't change. But, if the todo was sitting in the inbox
it would appear as a project under the inbox instead of being under its
collection.

So we simply catch this case and trigger a reparent. Of course comes
with a unit test.