Avoid rather nasty double free crash.
The crash was caused by m_strategy->onNodeRemoval(id) triggering another
call of removeNode with the same id, resulting in the original call to
removeNode trying to free the same pointer again.
By removing the pointer from the map early enough we avoid
the nested call to free the pointer as well.

Smart pointers could really help to detect stuff like this.