Simplify distribution of selection-models.
Instead of relying on the proper call order and setting the selection model in the modelstack,
we replace each usage of the model by a getter, and create the model already properly connected on demand.

This way we not only get rid of call order problems, but we also simplify passing around all the models to the different views.
The new approach therefore simplifies adding new views.