Separate monitor from pimitem.
The whole "cache the content of the akonadi item" idea was based on early misconceptions and would just result in a lot of unnecessary copying of data.
This patch instead gets the data directly from the akonadi item, and thus avoids the unnecessary copies.
Further the monitor part is separated from the items (as it's only used in the editor anyways).
As the PimItem is now really just a wrapper around Akonadi::Item, it doesn't have to be a QObject any longer (memorymanagement by QSharedPointer).

Overall this should simplyfiy PimItem and make it more lightweight to instanciate.