Simplify PimItemModel
There's a lot of commented out or unused features in there, just remove
them and make PimItemModel as simple as possible again. The only
exceptions are the presence of the date column (to make it convenient
to sort until we get a proper sort bar) and the support for PimItemRole.