IncidenceItem::isProject() is now more clever
Now it queries automatically PimItemServices and if the pim structure
cache is available it also determines if an item is a project based on
that. It simplifies a bit the logic in TodoMetaDataModel then.

It pushes IncidenceItem a bit higher in the dependency graph, but I
think it's the way to go.