Call the replace handlers on task update
It is not strictly necessary on the semantic level as the object didn't
get replaced, it got just updated. That said, it is convenient on the
outside to call the replace handlers nonetheless as it gives a mean to
be notified on any changes in the object, not having to monitor all its
signals which are sometimes too fine grained.