Add ApplicationComponents which allows to build widget based shells
This one wires the behavior between presentation model and widgets. It
is not a widget itself as we want to use it in the standalone
application and the kontact part where the building blocks will be put
together differently.

Also change a bit the main application to verify it works.