Force a full sync on the knut resource before launching cucumber
For some reason, since the big code style change in akonadiserver, the
behavior changed. On startup the knut resource is not fully synced.
Most notably if you attach akonadiconsole you notice that the attributes
on collections are missing, and all collections are reported empty.
After a full sync, the attributes are present and statistics in the
collections are correct (even though akonadiconsole still doesn't list
any items).

Hopefully should make us green on the CI again.

CCMAIL: dvratil@redhat.com