Port TaskQueries implementation to LiveQuery
Needed to adjust a bit the tests because we still mock too much. Use the
opportunity for unmocking a bit.

Also introduce a small regression in
shouldRemoveParentNodeAndMoveChildrenInTopLevelResult which should be
fine as that's a case the user couldn't see anyway. It was mostly a
workaround which got it to work anyway, a proper fix will come in time
once we get a proper cache system.