Domain change: Topic goes away in favor of Tag
After discussing with Björn Balazs, it looks like the way out of our
doubts regarding Contexts + Topics vs Tags is to go with Contexts +
Tags.

So remove everything related to Topic, and maps the new Tag type to
akonadi tags of type PLAIN. They will be organized along a flat list and
can be associated to any type of artifact (topics were limited to
notes).