Deal with the fact that search job returns any collection which matches
In the case where we're looking for children of a particular nore, we
might get back collections coming from other branches. Those ones will
never have the root we're looking for as ancestor, so stop climbing up
the tree at the collections root too. Otherwise it was looping
infinitely in such a case. Now it is fixed.