Implementation of associate : itemChanged is now taking into account by tags
issue : children of a newly associated task are moved too, that's okay but they
don't hold the tag, they're here because of our implementation of the hierarchy.
This lead to a weird behavior : if you associate a task to a tag when the parent
of this tag is already connected, it makes the parent and the child on the same level.

I guess we should avoid the user to associate a task to a tag when the parent
is already linked to this tag. Sorry for the speech :)

REVIEW: 120873