Switch to the new MockObject in our tests
Because of that change, we have to switch to the use of shared pointers
for queries, repositories and internal interfaces in the akonadi module.

That simplifies quite a bit the code in the end, we can finally get rid
of most of the destructors and the horrible "own interfaces" flags which
were introduced because of the old mock_object use.