Akonadistorage Unit testing item dissociation from tag
I add a new item in testdata that is linked to philosophy-tag, I' retrieve
the item, check that it is linked to the tag and dissociate it. I ask for
updating the item to akonadi and I verify that that the updated item is
not related to the previous tag I just dissociated from.