dissociateAll() implementation within akonaditaskrepository, unwip feature
Summary:
Since we use a Akonadi::Item* to clearItem we are oblige to mock this with
any() on mockitopp-side. Fixing the test of availablepages model due to
changes on the method call dissociateAll replacing dissociate. Fixing
workday acceptance test, "Buy apples" was changed by my newly created
acceptance test for tag reset so I changed it to "Buy kiwis" and make it
more resiliant at the same time, by checking it wasnt already in the workday
in the first place.

adapting workday acceptance test

Reviewers: ervin, bensi, remibenoit

Differential Revision: https://phabricator.kde.org/D84