Fixing monitor implementation to catch ItemsTagsChanged signals
Summary:
Akonadi emit a itemsTagsChanged when a tag is updated, so we have a list of
related items, we so far were only connected to itemChanged so we were missing
the items dissociated from a tag that got deleted. In order to fix this we
now emit manually a signal of itemChanged on each item that have concern on
the tag update (change or delition).

Test Plan: akonadistorage test is enough

Reviewers: bensi, remibenoit, theovaucher, ervin

Reviewed By: ervin

Differential Revision: https://phabricator.kde.org/D90