Add AkonadiFakeStorage
Summary:
This is an in-memory implementation of the Akonadi server behaviors we
depend on. It passes the same test suite than our implementation talking
to a real server.

This required to improve AkonadiFakeData returned values. We make them
look closer to the real system, especially by reconstructing the
ancestors chains for items and collections.

We also provide a couple of helpers to know which is the highest
allocated id for each of the value types (items, tags, collections).

Reviewers: bensi, franckarrecot, remibenoit

Differential Revision: https://phabricator.kde.org/D176