QueryTreeModel is now allowing DnD on the empty part by default (non valid index)
Summary:
By default it's possible, however the QueryTreeModel trying to be generic
is making possible to change this through setRootIndexFlag. So in our case
we use the default one, but it still can be changed.
QueryTreeModelTest was already under the assumption that this mechanism
was possible even though it was not (because of the Qt::NoItemFlags that's
why there are no test added)

Reviewers: remibenoit, bensi, ervin

Differential Revision: https://phabricator.kde.org/D118