Fix the cppcheck warnings around ctors
Summary:
It seems that the cppcheck I have here is more clever than the one on
the CI (I run 1.69 here) and as such catches more style issues. Most
notably it complains about missing explicit for ctors.

Fix all the ones which got found and use a cppcheck-suppress line for
the false positive missing ctors or the ctors which are meant to be
implicit anyway.

Reviewers: bensi, franckarrecot, remibenoit

Differential Revision: https://phabricator.kde.org/D254