Don't loose the parent collection during collection changes
Summary:
Sometimes when modifying a collection we have it only partially, in
particular its parent collection might not be set, in such a case we
should keep the last known parent collection inside of the data set.

Previously we just lost that piece of information which causes issue for
full stack testing.

Reviewers: bensi, franckarrecot, remibenoit

Differential Revision: https://phabricator.kde.org/D259