Don't loose the parent collection during item changes
Summary:
Sometimes when modifying an item we have it only partially, in
particular its parent collection might not be set, in such a case we
should keep the last known parent collection inside of the data set.

Previously we just lost that piece of information which causes issue
for full stack testing.

Reviewers: bensi, franckarrecot, remibenoit

Differential Revision: https://phabricator.kde.org/D260