Allow to run the acceptance tests on fake data
Summary:
Now we can run the acceptance tests either within akonaditest like
before or using the akonadi fake data. That basically cuts the execution
time in slightly more than half.

We got now two targets:
 * make cucumber-run-done, runs with fake data;
 * make cucumber-akonadi-run-done, runs with akonaditest.

During "make test" the cucumber-run-done is the new default it will be
both better for developers and avoid spending too much CI resources. The
old targets with akonaditest are kept around in case we need to compare
the behavior in the future. Intent still is to catch the behavior
differences between the fake data and the real akonadi using the
integration tests of the fake data unit tests.

Reviewers: bensi, franckarrecot, remibenoit

Differential Revision: https://phabricator.kde.org/D262