Simplify the behavior of the live queries
Summary:
Right now we got quite some duplication of code in lambdas for the
behavior of the live queries. So instead of specifying all the behavior
per query we factor out the convert/update/represents part at the query
creation point.

This is a first step to factor more code around the queries, we can see
quite some similarity emerge already with that commit.

Reviewers: bensi, franckarrecot, remibenoit

Differential Revision: https://phabricator.kde.org/D266