Have TaskInboxPageModel deal only with tasks
Summary:
It is now using the new task specific inbox query. It also required
adjusting the acceptance tests as the notes disappeared from the inbox
and a tagged task is now listed there (all due to the new definition).

Also had to remove project dnd test related to notes since we couldn't
find a note to drag anymore.

Reviewers: bensi, franckarrecot

Differential Revision: https://phabricator.kde.org/D485