Allow to specify the content types of Akonadi::DataSourceQueries
Summary:
With this commit, an Akonadi::DataSourceQueries can be specialized for a
given family of collections. At construction time we can decide for it
to list and search only for task sources, or note sources or both.

Also adjust the dependencies of both Renku and Zanshin so that one lists
only note related sources and the other one only task related sources.

This effectively complete the Zanshin/Renku split: one is pure tasks,
the other pure notes.

Reviewers: franckarrecot, bensi

Differential Revision: https://phabricator.kde.org/D501