Allow to control the default data source from DataSourceQueries
Summary:
Now that it is the place where we control the type of data sources we're
interested in, no need to do that in the note or task repositories. It
is a first step toward centralizing that feature.

Note the ugly notifier object and corresponding hacks, that is necessary
to allow mockitopp to do its job without crashes. Inheriting directly
from QObject doesn't work with it.

Reviewers: bensi, franckarrecot

Differential Revision: https://phabricator.kde.org/D510