Repair the delegation mechanism
Summary:
We have to set a valid organizer for the emails to go through so we
first ask for it. We also have to make sure a properly loaded calendar
is set on the ITIPHandler before use. It claims it will initialize one
on demand but that's just not true in our case.

Reviewers: bensi, franckarrecot

Differential Revision: https://phabricator.kde.org/D593