Workaround broken fetch items based on their tag
Summary:
It turns out that FetchItemJob for tag contexts doesn't list us the
tagged items properly... until the tagged item has been fetched by
another unrelated job. It obviously create weird issues which are hard
to reproduce in tests (I didn't manage to isolate a unit test at least).

It seems to be fixed in akonadi/qt5 if I trust git log there.

For the time being, work around that issue with a super-inefficient
implementation in our help functions. We fetch all items and trash the
ones which don't have the tag we're looking for.

Clearly something to remove and reevaluate once we port to Qt5.

Reviewers: bensi, franckarrecot, dvratil

Differential Revision: https://phabricator.kde.org/D652

CCMAIL: dvratil@kde.org