Kill all the job handlers when ApplicationModel is gone
Summary:
This way we make sure no unwanted function will be triggered once the
user decided to exit the application. This is admittedly a cheap fix for
the "crash on exit" we're seeing from time to time. A more comprehensive
fix would be to audit all out lambdas which tend to capture this a bit
too often for my taste. That said, even if we did this long auditing,
our code would still potentially execute code and trigger fetches for
nothing at exit, so let's avoid that.

Reviewers: bensi, franckarrecot

Differential Revision: https://phabricator.kde.org/D670