QueryTreeModelBase prevents cycle creation
Summary:
For internal drag and drop, we forbid the user to drop an item on one of
its descendents. Otherwise it might create cycles in the data leading to
crashes.

Ideally the check on the data should be done at a lower level, but that
requires the caching facilities which are still in the work.

Reviewers: bensi, franckarrecot, dfaure

Differential Revision: https://phabricator.kde.org/D672