Workaround KMime trimming the last empty line
Summary:
That should fix the jumping cursor issue we see in Renku. It is due to
the editor sending to save the latest empty line, then the change being
signalled back, the empty line removed... then the line on which the
cursor was is gone, making it jump at the beginning of the editor.

It seems that KMime does other weird stuff with empty lines, but that
was the most annoying one for us so far.

Reviewers: bensi, franckarrecot

Differential Revision: https://phabricator.kde.org/D677