Have Escape work from anywhere
Summary:
The action mapped on escape was scoped fairly strictly to only the
filter and quick add widgets. That was mainly to make sure that escape
would have a different behavior in the central view (clear the
selection). Alas, we'd like the action behavior everywhere in the
application so that escape always gets you back to the central view.

The solution is simply to have the action triggerable from anywhere, but
to make sure it's properly enabled/disabled when the central view
gets/looses the focus.

Reviewers: bensi, franckarrecot, dfaure

Differential Revision: https://phabricator.kde.org/D678