Always expandAll() in the central view
Summary:
Unlike the sources view, the central view has a filter proxy, so we
might not always the row inserts when grand children were already
loaded. It's generally not a big issue, except in this view where one
can't expand/collapse nodes. So always expand totally to be on the safe
side.

Reviewers: bensi, franckarrecot

Differential Revision: https://phabricator.kde.org/D692