Make the editor more resilient to spurious trims
Summary:
It turns out that not only KMime trims the text or the title on some
condition. There's apparently some weird interaction I can't pin point
between akonadi and kcalcore serialization formats which lead to space
trims at the end of title or text.

So instead of purely comparing the last known strings and the new ones,
we are more lax with a regexp allowing whitespaces differences at the
end of title and the end of text.

Reviewers: bensi, franckarrecot

Differential Revision: https://phabricator.kde.org/D693