On shutdown make sure the ApplicationModel goes first
Summary:
This is necessary when everything is integrated, on shutdown lists are
emptied which makes the GUI react and ask for more information which is
unwanted at that point. Make sure we cut the links between model and GUI
before the GUI is destroyed further.

This fixes a crash on exit which showed up if you had a project selected
for instance and closed the window.

Reviewers: bensi, franckarrecot

Differential Revision: https://phabricator.kde.org/D711