Children of unlisted collections are now listed
Summary:
This is a behavior change in the akonadi server. Previously, children of
unlisted (!enabled && !referenced) collections were not fetched in a
recursive fetch. This is not the case anymore.

In fact I think this new behavior is much better, so adjusting the test
and our fake implementation instead.

CCMAIL: dvratil@kde.org

Reviewers: bensi, franckarrecot, dfaure, mlaurent, dvratil

Differential Revision: https://phabricator.kde.org/D878