Replace the flaky features run script with a proper executable
Summary:
This way we can properly wait for the cuke steps server to show up
before launching cucumber which removes the race (and thus random
failure) we were seeing on the acceptance tests.

I had to adjust cucumber-cpp a bit so that it doesn't quit after the
first connection attempt.

Reviewers: bensi, franckarrecot, dfaure

Reviewed By: dfaure

Differential Revision: https://phabricator.kde.org/D952