Move the notifier to a global static
Summary:
This is necessary to fix a leak when the DataSourceQueries interface is
used through MockObject. We don't go through ctor/dtor in such cases as
we're supposed to have pure interfaces with no data. The notifier in
fact doesn't need to be held by every instances of that interface so
share it using a global static.

Reviewers: bensi, franckarrecot, dfaure

Differential Revision: https://phabricator.kde.org/D1064