Expand the row when selected and show more information
Summary:
Added some more information to display in the item delegate of the
central view. Now we display the due date right aligned if available and
also we show the first few lines of the artifact text when the item is
the current item (requiring a higher row).

It led to a massive refactoring of the delegate.

Reviewers: bensi, dfaure, franckarrecot

Reviewed By: franckarrecot

Differential Revision: https://phabricator.kde.org/D2738