Turns out callsigns are gone
It prevents detecting properly on the remote URL of the repository if we
set that option. So let's remove it. In the case of David not using
anongit as pull remote... either change your remote or see with Ben what
could be done about it, I'm out of options now. In the meantime remember
to manually set the repository on your reviews.

CCMAIL: faure@kde.org