Zanshin: remove code that triggers the 0.2 data migration.
Summary:
When using zanshin on a new computer, the local zanshinrc is empty so
the migration happens even if it is not needed. This leads to tasks with
children becoming projects, unexpectedly.

If a different migration is needed again in the future, we have to find
a way to store the version into the items themselves, rather than locally.

If someone still hasn't upgraded from 0.2, they can run the migrator manually
when they do.

Reviewers: ervin

Reviewed By: ervin

Subscribers: #zanshin

Differential Revision: https://phabricator.kde.org/D3767