New feature: running task widget
Summary:
Start a task and a small line appears at the top of your desktop.
Moving the mouse onto that line expands to a bar that shows
      [Stop]  Task title   [Done],
as a reminder of what the current task is.

Design (new classes) :
    - RunningTaskModel to store the running task (with base interface for type safety of stubs).
    - RunningTaskWidget and PageView on top of that model
    - TaskApplicationModel to create the RunningTaskModel with its deps
    - TaskApplicationComponents to connect PageView and RunningTaskWidget to RunningTaskModel

Test Plan:
No cucumber tests because OpenSuSE botched the cucumber packaging.
    (https://bugzilla.opensuse.org/show_bug.cgi?id=1015940)
Unit tests for all new classes though, and for serialization of the 'running' attribute.

Reviewers: franckarrecot, ervin

Reviewed By: ervin

Subscribers: #zanshin

Differential Revision: https://phabricator.kde.org/D3032