Remove the unused list/unlist/bookmark collection feature
Summary:
This was introduced during 0.3 times for a specific Munich request. Was
never used otherwise and even unreachable from the GUI by the user
except when a specific flag was set. Needless to say it had plenty of
time to rot and break without us noticing. Time to clean it up.

Reviewers: #zanshin, dfaure, bensi, franckarrecot

Differential Revision: https://phabricator.kde.org/D5830