Add a CachingStorage
Summary:
This one will forward all requests to the real storage but will also
store the results in the cache. For now it only forwards.

In this patch we plug the CachingStorage instead of the regular
Storage in the acceptance tests and the applications, we also add a new
set of integration tests going through this new storage. This should
help catch regression early.

The behavior parts to feed the cache will come in later commits.

Reviewers: #zanshin, dfaure, bensi, franckarrecot

Differential Revision: https://phabricator.kde.org/D5831