Fix lack of confirmation when deleting an item with hidden children.
Summary: We need to go up to the source model to see if an item has children.

Test Plan:
In a project, add a child to a task, and give it a start date in the future.
Without "Show future items" enabled, the child disappears as expected.
Now try deleting the parent task, it used to delete it without confirmation as if it
had no children, which was very dangerous (data loss of the invisible children).

Now at least it warns, so the user can wonder why this talks about children when none
are visible, and can enable the showing of future items to find out what they are
before possibly attempting another deletion.

Reviewers: ervin

Reviewed By: ervin

Subscribers: #zanshin

Differential Revision: https://phabricator.kde.org/D12636