Explicitly request TagAttribute in Akonadi Monitor's tag fetch scope
Summary:
Akonadi errornously returns all Tag attributes by default when queried
via TagFetchJob, which Akonadi Monitor used to do internally. This has
changed in the recent Notification Payload feature when the Tag is sent
directly as part of the change notification. This uses a slightly different
code path on the serve to assemble the Tag as part of the tag change
notification. This code path honors the TagFetchScope settings and only
returns tags that were explicitly requested, thus breaking your test,
which did not ask for the attribute explicitly.

Test Plan: Fixes akonadistoragetest

Reviewers: ervin

Differential Revision: https://phabricator.kde.org/D14877