Fix workdaystartdateeditfeature and workdayduedateeditfeature.
Summary:
Changing a property (e.g. start date) in the editor doesn't have
immediate effect, save() is called by a timer. waitForStableState()
waits for jobs and THEN waits in the event loop, but if that event loop
creates more jobs, it's not waiting for those. Better save right away
when going to a different page.

Test Plan: ctest -R dateeditfeature

Reviewers: ervin

Reviewed By: ervin

Subscribers: #zanshin

Differential Revision: https://phabricator.kde.org/D18423