Fix context not disappearing from the treeview upon removal.
Summary:
It turns out that Akonadi emits itemRemoved without any payload
in the item, while AkonadiFakeData was emitting full items.

So my wrong implementation in the LiveQueryIntegrator (which required
a payload) was working with fake data but not with real Akonadi.

Just comparing the item ID like we do for tasks and projects is much
simpler anyway.

Reviewers: ervin

Reviewed By: ervin

Differential Revision: https://phabricator.kde.org/D19147