Add a migrator for converting tags to context items
Summary:
This migrator creates one context item for each context tag,
then associates the tasks with those contexts.

The tags are just left there. This makes a temporary downgrade possible,
as well as making it possible to rerun the migrator if a bug is found and fixed.

Test Plan: New unittest passes.

Reviewers: ervin

Reviewed By: ervin

Differential Revision: https://phabricator.kde.org/D19195