Fix uids for contexts created by the migrator
Summary:
A tag's GID isn't a good UID for a todo, the caldav server rejects
UIDs with spaces.

So do this the right way instead: let akonadi generate a proper UID
for the context, and then point to that from the tasks.

Test Plan: unittest still passes, now I'll convert my tasks again...

Reviewers: ervin

Reviewed By: ervin

Differential Revision: https://phabricator.kde.org/D19217