Fix interoperability issue with kolab server which uppercases all properties
Summary:
iCal says property names are case insensitive, but kcalcore doesn't
implement that. 99% of kdepim uses uppercase property names for custom
properties, which is actually consistent with other iCal properties.

Creating a project or a context and then syncing with the kolab server
would lose the fact that it was a project or context, forcing the user
to change the type of the item again (easy for a project, impossible for a context).

I looked a bit into fixing kcalcore's CustomProperties but calling .toUpper()
everywhere sounds slow, uppercasing the keys could in theory break users of
the QMap accessor, and more importantly, is there anyone to review such a change? ;)
OTOH this commit is how the code should have been in the first place, to
be consistent with the rest of kdepim.

 This commit includes automatic migration from "Project" to "ISPROJECT"
  so that the change is transparent to users.

Test Plan:
Newly created contexts (by the migrator or by hand) no longer disappear
after a caldav sync.

Reviewers: ervin

Reviewed By: ervin

Differential Revision: https://phabricator.kde.org/D19238