Fix findTopLevel() being the same as findInboxTopLevel()
Summary:
findInboxTopLevel() returns tasks without a parent (neither a
parent task nor a project), as expected.

findTopLevel() was doing exactly the same, and isn't used anywhere
except in unittests, currently.

But I'm going to use it for an "All Tasks" page, where we do want tasks
that belong to a project, but where child tasks (of other tasks)
shouldn't be returned since they'll be visible under their parent task.

Port most unittests to findInboxTopLevel so they don't require a cache.

Test Plan:
test passes. Also tested in the app together with a WIP commit
for adding an "All Tasks" page.

Reviewers: ervin

Reviewed By: ervin

Subscribers: #zanshin

Differential Revision: https://phabricator.kde.org/D19819