New page: all tasks
Summary:
This page shows all tasks. It's especially useful to search for a task
that one forgot in which project it was classified.

Test Plan:
Tests pass; zanshin shows new All Tasks page.
Adding and removing works.
One problem left: shall we disable Promote? The page shows no projects
so I think it's the wrong place for that, the task would just disappear.

Reviewers: ervin

Reviewed By: ervin

Subscribers: #zanshin

Differential Revision: https://phabricator.kde.org/D19957