Remove stylesheet in PageView so it reacts to colors and fonts changes
Summary:
The stylesheet was added to hide what looks like an expand/close indicator
on a tree that is always expanded (setItemsExpandable(false)).

On the other hand, the use of a stylesheet at all introduced these two
bugs: changing the font globally (kcmshell5 fonts) or the colors
(kcmshell5 color) would work on most of zanshin EXCEPT on the treeview
in the middle, which kept the old palette or the old font. Awful.

Redone with a proxy style to avoid these issues.

Test Plan: kcmshell5 colors now applies the colors to all of zanshin

Reviewers: ervin

Reviewed By: ervin

Differential Revision: https://phabricator.kde.org/D22578